- added build/war/WEB-INF/classes to the ddl.classpath, since it is (now) needed in order to get at the UserType for the vocabulary columns.
- removed clean_persistence as a dependency on ddl, due to a subtle problem: the
determination of whether hbm_xml needs to be run occurs *before* the ddl rule gets called.
so the build determines that hbm_xml does not need to run; deletes the resources created
by hbm_xml with a call to the ddl target; then skips the call to hbm_xml and goes on to
generate the ddl, which fails without the hbm_xml resources.

git-svn-id: https://forge.abcd.harvard.edu/svn/screensaver/trunk@137 80c535d0-b110-0410-b4de-9279f94142c3