continued progress on the visits portion of the schema. at this point, inheritance is working. still need to flesh out relationships between the different entities.
(the prior two checkins were supposed to be on a branch, but landed on the trunk due to
my mistake. the log messages are messages about saving partial work to a branch. this
partial work was the early stages of this checkin: getting Visit and its two subclasses
to do inheritance right in terms of Hibernate.)

git-svn-id: https://forge.abcd.harvard.edu/svn/screensaver/trunk@297 80c535d0-b110-0410-b4de-9279f94142c3