from dialog for RT #55975:
Andrew and I discussed
this this afternoon, and seemed to come up with a workable solution:

- no code assumes wells have been created
- code assumes that if any well for a library has been created, then all
wells for the library (384 * numPlates) have been created
- if code needs a well that is not there, it creates all the wells for
the library the well is in
- code creates wells with wellType.EMPTY, unless it knows better. it
likewise changes wellType of existing wells when it knows better. (the
library loader will make wells experimental, and the NENEWTDFImporter
will make wells Buffer or whatever.)

Once this is done, I'll leave it to Andrew to make sure the assay well
types are in agreement with the library well type. I should check in my
part tomorrow.


git-svn-id: https://forge.abcd.harvard.edu/svn/screensaver/trunk@887 80c535d0-b110-0410-b4de-9279f94142c3