- created screensaver.properties file for build-time and run-time configuration - ScreensaverProperties classes mirrors the contents of the .props file - used screensaver.properties to give location of the executable `babel`
this is an initial version. i intend to extend this by modifying the deploy scripts in flotsam_jetsam/orch-bin
to acquire the right properties file. checking in code from my desktop to transfer development over to
orchestra.

i also intend to try to use the screensaver.props file for selecting the appropriate log4j.props file, and
possibly the datasource.props file as well. but first i am going to nail the `babel` issue.

git-svn-id: https://forge.abcd.harvard.edu/svn/screensaver/trunk@908 80c535d0-b110-0410-b4de-9279f94142c3