made AbaseTestset properties testsetDate and comments NOT NULL.
motivation: adding comments to the business key, but leaving it NULL, caused problems in where
the AbaseTestset 2-arg constructor was adding the testset to the set of testsets for the
screen. because comments wasnt set yet, i had a problem where the hashCode() for the testset
changed after it was added to the set. then later on, i got dangling testsets that were not
getting deleted properly. i guess i only needed to make comments NOT NULL, but i saw in
ScreenDB that the date was also always set, so I changed that too.

git-svn-id: https://forge.abcd.harvard.edu/svn/screensaver/trunk@1124 80c535d0-b110-0410-b4de-9279f94142c3