fixing logging disaster in two steps:
first, instead of using a Spring PropertyPlaceholderConfigurer to select the log4j props file,
revert to the previous way of selecting the logging file, which is just copying over the log4j
props file specified in screensaver.props to the correct place in the distro;

second, creating an ant target for an Eclipse Builder that does the same copy-over in
.eclipse.classes. this first step fixes logging for all tomcat launches, and the second step
makes logging work well for apps launched from the Eclipse Ap Launcher.

essentially, this solution reverses the concept behind the previous broken solution: now, we
make the log4j props file selection mechanism work great for external deploys, and figure out
a way to make it work for Eclipse Application Launcher after that. this is actually much cleaner,
thanks to Eclipse being so damned plugin friendly.

i put a comment on the new build.xml rule to describe how to set up the Eclipse Builder. setting
it up dropped a .eclipseBuilder dir in my project directory, which i could check in instead, if
everyone on the team thinks its a cool enough solution to be a project default.

git-svn-id: https://forge.abcd.harvard.edu/svn/screensaver/trunk@1356 80c535d0-b110-0410-b4de-9279f94142c3