modified some entity model cascades:
compound => well from "all" to "save-update", since don't want to delete a well just because the compound was deleted.

well => compound from "all-delete-orphan" to "save-update", since this apparently didnt have the semantics i was intending: the delete was cascading from well to compound even though the compound was in other wells! (i think the delete-orphan is really intended for one-to-many, not 
many-to-many)

well => silencing reagents from "all-delete-orphan" to "save-update" for the same reason as
well => compound.

fixes rt#81927

git-svn-id: https://forge.abcd.harvard.edu/svn/screensaver/trunk@1462 80c535d0-b110-0410-b4de-9279f94142c3