- Added WellVolumeAdjustment entity, factored out of LabCherryPick. LabCherryPick now has a set of WellVolumeAdjustments, though the current Java implementation only allows for a single WellVolumeAdjustment per LabCherryPick.  A WellVolumeAdjustment can alternately be associated with a WellVolumeCorrectionActivity (see next).
- Reconfigured the ScreeningRoomActivity entity hierarchy.  Added new
top-level parent entity, Activity.  Activity is now the parent of
ScreeningRoomActivity and a new AdministrativeActivity entity.  Activity
encapsulates who/when/what, while AdministrativeActivity adds who/when
approved the activity. Also added WellVolumeCorrectionActivity, an
AdministrativeActivity subtype, to track misc adjustments to library copy well
volumes. 

- AdministratorUser now has a to-many relationship with
AdministrativeActivity.

- Nearly resolved #79216 (data model support for library copy well volume
adjustments).  To complete this ticket, we need to migrate the
copyInfo.microliterWellVolumes to
wellVolumeCorrectionActivity.wellVolumeAdjustments, and have 
CherryPickRequestAllocator instead inspect/respect these WVAs; but this is 
best done in a separate commit.

- f+j/manual_schema_updates/1521.sql takes care of the (significant) schema 
migration required for this update.  It also replaces screensaver_revision 
with a new schema_history table.

- Moved assorted VOLUME_SCALE constants to Well.VOLUME_SCALE.


git-svn-id: https://forge.abcd.harvard.edu/svn/screensaver/trunk@1521 80c535d0-b110-0410-b4de-9279f94142c3