added code to synchronize RNAiCherryPickScreenings. These are very similar to LibraryScreenings, and I managed to reuse a lot of code by extracting LibraryScreeningSynchronization code into an abstract superclass, and made the RNAiCherryPickScreeningSynchronizer inherit from there. i also extracted the code for determining volume transferred per well into its own class, tho it is a bit of a mystery to me why i did that. oh well, its a quaint, short little class.
about the RNAiCherryPickScreenings - i need the ScreenDB 'RNAi Cherry Pick Screening' visits to
have a non-null CPR number (visits.cpr_number_for_cp_screen) in order to do this job properly. i
have discovered that 20/30 such visits do not have a CPR number. i asked Katrina about it and i
imagine she well be able to just fill some of them in, but it won't always be so easy because these
cherry pick requests are no longer being entered into ScreenDB. so i might have to write a little
script to back-populate these, which in turn might lead to further changes to the
RNAiCherryPickScreeningSynchronizer to work this all out.

git-svn-id: https://forge.abcd.harvard.edu/svn/screensaver/trunk@1596 80c535d0-b110-0410-b4de-9279f94142c3