- Resolved #89651 (ScreenResultParser OutOfMemory on LSF cluster).
- Created analog of io.workbook package in io.workbook2 package, using
the JExcelAPI library ("jxl") as the underlying technology.  This
library is considerably less memory-intensive than the POI/HSSF
library, allowing us to parse 50MB+ Excel files with only a few
hundred MBs of heap (still a crazy ratio, but...).  Classes in these two 
parallel packages are similar, but not worth sharing code between them, 
as the intention is to get rod of io.workbook (i.e. POI/HSSF) altogether.  

- Upgraded jxl.jar from 2.6.3 to 2.6.5.  Appears to fix a date parsing
bug.  

- ScreenResultParser now uses io.workbook2 classes.  

- Renamed 'metadata' occurrences in ScreenResultParser to 'detaHeader'
(was a legacy naming) 

- ParseErrorManager now only creates an annotated error workbook upon
request.  This improves time and memory performance.

- Updated ScreenResultParserTest and test excel files, as necessary

- Added JExcelApiMemoryTest and POIMemoryTest classes.  Will probably
remove, but wanted to save in VC repository.


git-svn-id: https://forge.abcd.harvard.edu/svn/screensaver/trunk@1669 80c535d0-b110-0410-b4de-9279f94142c3