- ScreenResultViewer data table now uses dataTableNavigator.jspf.  So now all data tables in Screensaver now use dataTableNavigator.jspf, and newsViewer.jsp proudly notes this fact.
- Resolved #99596 ("screen result data table # rows")

- Within ScreenResultViewer code (i.e., multiple classes), removed all calls
to DataTableRowsPerPageUISelectOneBean.setAllRowsValue().  Was causing too
many bugs w.r.t the timing of calculating the "all rows" count for the various
ScreenResultDataTable subclasses.  As it turns out, if we don't set this at
all, the value of DataTableRowsPerPageUISelectOneBean.getSelection() is null
when the user selects the "all" option, and null happens to be the value that
tells our ScreenResultDAO methods to retrieve all rows!  I'm not proud of this
tenuous coding practice, but damn it, it works, and I'm keeping it!


git-svn-id: https://forge.abcd.harvard.edu/svn/screensaver/trunk@1963 80c535d0-b110-0410-b4de-9279f94142c3