[This revision is not ready for production, as Well.molfile is disabled, ReagentSearchResults needs to be optimized, and more testing need to be performed.  It is fully functional, however.]
- Partial resolution of #97774 ("Well, Reagent, Vendor, AnnotationValue entities & relns").
Added Reagent entity.  Currently, just a placeholder entity (only has an ID)
to allow for proper data modeling.  The code was getting way too cumbersome,
trying to use Well as an implicit container for annotation values; in
particular, ReagentSearchResults was in need of too many (inefficient) DAO
methods.  In the future we will have Reagent manage a well's compounds and
silencing reagents (per #102278).

- ReagentVendorIdentifier is used as the entity ID of Reagent, via
@Embeddable.

- Added AnnotationValue many-to-one relationship with Reagent.  This allows
entity model-based access to a Reagent's annotation values, which had to be
acquired via DAO methods previously.  Also allows simpler
ReagentSearchResults, which was parameterized on Well previously (as a hack),
but is now be parameterized on Reagent; no longer calls DAO methods to obtain
annotation values for reagents/wells.

- Added Well many-to-one relationship with Reagent.

- ReagentViewer is now Reagent-based, rather than Well-based (a hack,
previously).

- Library contents importers now create Reagents on-demand, as necessary.

- In StudyViewer, replaced AnnotationViewer with ReagentSearchResults.
Deleted AnnotationViewer and its supporting data table classes.

- Consolidated Library.createWell() methods into a single method (affecting
mostly just test code).

- Removed 'parentWellOfInterest' property in GeneViewer and CompoundViewer,
since it was only used in displaying a message to the user.

- Updated/fixed unit tests.

- Added ScreenCreator, which was intended for earlier rev.


git-svn-id: https://forge.abcd.harvard.edu/svn/screensaver/trunk@1990 80c535d0-b110-0410-b4de-9279f94142c3