- gather compounds for upgrading based on whether the set of pubchem cids is empty - removed usage of entity props Compound.isPubhcmeCidListUpgrader{Successful,Failed}. these stopped being meaningful, and at this point a failure is defined as the compound having an empty set of pubchem cids
the usefulness of the success/failure flags was much more important when the upgrader was not
running on the lsf queue, and prone to need to be restarted.

in preparation for removing the Compound.pubhcmeUpgrader properties. i may regret this later,
but probably not, and they are really ugly fields to have in such a high-profile entity

git-svn-id: https://forge.abcd.harvard.edu/svn/screensaver/trunk@2089 80c535d0-b110-0410-b4de-9279f94142c3