- Resolved #113119 (screen result importer terminates prematurely 449). Chalked it up to excessive memory usage, by both JVM and postgres.
- ScreenResultImporter and ScreenResultParser now support incremental
importing of screen result files, by plate ranges.  This allows the importer
to handle a "manageable" size of data, without running out of memory, or
overloading postgres (which otherwise is being asked to complete the entire
import in a single transaction).

- Rewrote LibrariesDAOImpl.loadOrCreateLibraryWells().  I no longer understood
the comments we had in there, and wanted to have the code be written in a way
that made it very clear of its behavior w.r.t the Hibernate session; that is,
how the Library passed in is used when it's in either a transient, managed, or
persisted state.  The impl no longer relies upon catching of a Hibernate
exception to deal with a transient library object, avoids removed use of HQL.

- ResultValueType.resultValues map changed from Map<Well,ResultValue> to
Map<WellKey,ResultValue>.  This avoids having to have a full Well object
available to index into the map, which should optimize code that accesses this
map.

- WellKey is now an @Embeddable, so that it can be used as key type of
RVT.resultValues map.

- ResultValue.resultValueId now uses a 'seqhilo' ID generator, to optimize
bulk insert speed.  Resolves #113472 (ResultValue insert optimization).

- Updated unit tests.

git-svn-id: https://forge.abcd.harvard.edu/svn/screensaver/trunk@2105 80c535d0-b110-0410-b4de-9279f94142c3