- Resolved #144055: screen search results refresh after screen edit. After adding a new screen immediately after login, the screen is now displayed in a properly initialized ScreensBrowser.
- ScreenViewer: Instead of using filtering to find and display a selected
entity in the EntitySearchResults, we find the entity in the already-loaded
search result.  This behavior fixed the above bug, but also provides more
intuitive behavior to a user that expects to be able to scroll through all
screens after finding a particular screen.
  
- ScreenDetailViewer notifies ScreenSearchResults to update itself after a
screen is edited or added.

- EntitySearchResults: renamed and simplified methods that scroll the data
table and methods that set entity to be viewed in entity view mode.


git-svn-id: https://forge.abcd.harvard.edu/svn/screensaver/trunk@2498 80c535d0-b110-0410-b4de-9279f94142c3