- Resolved #151579 (browse collaborators result set is wrong).
- 'My Collaborators' menu command is now 'My Associates', and opens a UserSearchResult
with the screener's lab head, fellow lab members and all screen associates (collabs, 
lab head, lead screener), as determined by SRU.getAssociatedUsers().

- WebDataAccessPolicy now restricts ScreeningRoomUser to "associated" users,
as determined by SRU.getAssociatedUsers().

- Added ScreenSearchResults.searchAssociatedUsers(), to provide an explicit mechanism 
for browsing this set of users.  Previously, we relied upon the data access policy to
simply restrict the search result to only associated users, but this is was both an 
inefficient and obtuse design.

- Fixed ScreeningRoomUser.getAssociatedUsers(), which was only returning
collaborators (was unused previously)

- Fixed bug in UserViewer that caused exception for ScreeningRoomUser login
accounts.


git-svn-id: https://forge.abcd.harvard.edu/svn/screensaver/trunk@2590 80c535d0-b110-0410-b4de-9279f94142c3